requirements.txt
flask
docker
paho_mqtt
criptography
meshtastic
